refacto(validator/test): split validator tests into multiple files
