+++ pkg/validator/validator_test.go
@@ -6,49 +6,19 @@ import (
 	"net/http"
 	"os"
 	"path/filepath"
-	"slices"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"gopkg.in/algolia/openvpn-auth-okta.v2/pkg/oktaApiAuth"
 	"gopkg.in/h2non/gock.v1"
 )
 
 const (
-	// used in TestLoadPinset
-	pin string = "SE4qe2vdD9tAegPwO79rMnZyhHvqj3i5g1c2HkyGUNE="
 	// used in TestWriteControlFile
 	controlFile  string = "../../testing/fixtures/validator/control_file"
 	oktaEndpoint string = "https://example.oktapreview.com"
 	token        string = "12345"
 )
 
-// used in TestReadConfigFile, TestLoadPinset
-type testCfgFile struct {
-	testName string
-	path     string
-	link     string
-	err      error
-}
-
-type testViaFile struct {
-	testName         string
-	path             string
-	usernameSuffix   string
-	expectedUsername string
-	expectedPassword string
-	err              error
-}
-
-type testEnvVar struct {
-	testName            string
-	usernameSuffix      string
-	allowUntrustedUsers bool
-	expectedTrusted     bool
-	expectedUsername    string
-	env                 map[string]string
-	err                 error
-}
 
 type testControlFile struct {
 	testName string
@@ -329,329 +299,6 @@ func TestSetup(t *testing.T) {
 	}
 }
 
-func TestCheckPasscode(t *testing.T) {
-	t.Run("Parse password with passcode", func(t *testing.T) {
-		setEnv(setupEnv)
-		v := NewOktaOpenVPNValidator()
-		_ = v.LoadEnvVars(nil)
-		v.api.UserConfig.Password = "password123456"
-		unsetEnv(setupEnv)
-		v.parsePassword()
-		assert.Equal(t, "password", v.api.UserConfig.Password)
-		assert.Equal(t, "123456", v.api.UserConfig.Passcode)
-	})
-}
-
-func TestReadConfigFile(t *testing.T) {
-	tests := []testCfgFile{
-		{
-			"Valid config file - success",
-			"../../testing/fixtures/validator/valid.ini",
-			"",
-			nil,
-		},
-		{
-			"Valid config file link - success",
-			"",
-			"../../testing/fixtures/validator/valid.ini",
-			nil,
-		},
-		{
-			"Invalid config file - failure",
-			"../../testing/fixtures/validator/invalid.ini",
-			"",
-			fmt.Errorf("Missing param Url or Token"),
-		},
-		{
-			"Invalid 2 config file - failure",
-			"../../testing/fixtures/validator/invalid2.ini",
-			"",
-			fmt.Errorf("key-value delimiter not found: UsernameSuffix\n"),
-		},
-		{
-			"Missing config file - failure",
-			"MISSING",
-			"",
-			fmt.Errorf("No ini file found"),
-		},
-		{
-			"Config file is a dir - failure",
-			"../../testing/fixtures/validator/",
-			"",
-			fmt.Errorf("No ini file found"),
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.testName, func(t *testing.T) {
-			v := NewOktaOpenVPNValidator()
-			v.configFile = test.path
-			if test.path == "" {
-				_ = os.Symlink(test.link, "api.ini")
-			}
-			err := v.ReadConfigFile()
-			if test.path == "" {
-				_ = os.Remove("api.ini")
-			}
-			if test.err == nil {
-				assert.Nil(t, err)
-			} else {
-				assert.Equal(t, test.err.Error(), err.Error())
-			}
-		})
-	}
-}
-
-func TestLoadPinset(t *testing.T) {
-	tests := []testCfgFile{
-		{
-			"Valid pinset file - success",
-			"../../testing/fixtures/validator/valid.cfg",
-			"",
-			nil,
-		},
-		{
-			"Valid pinset link - success",
-			"",
-			"../../testing/fixtures/validator/valid.cfg",
-			nil,
-		},
-		{
-			"Missing pinset file - failure",
-			"MISSING",
-			"",
-			fmt.Errorf("No pinset file found"),
-		},
-		{
-			"Pinset file is a dir - failure",
-			"../../testing/fixtures/validator/",
-			"",
-			fmt.Errorf("No pinset file found"),
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.testName, func(t *testing.T) {
-			v := NewOktaOpenVPNValidator()
-			v.api = oktaApiAuth.NewOktaApiAuth()
-			v.pinsetFile = test.path
-			if test.path == "" {
-				_ = os.Symlink(test.link, "pinset.cfg")
-			}
-			err := v.LoadPinset()
-			if test.path == "" {
-				_ = os.Remove("pinset.cfg")
-			}
-			if test.err == nil {
-				assert.Nil(t, err)
-				assert.True(t, slices.Contains(v.api.ApiConfig.AssertPin, pin))
-			} else {
-				assert.Equal(t, test.err.Error(), err.Error())
-			}
-		})
-	}
-}
-
-func TestLoadViaFile(t *testing.T) {
-	tests := []testViaFile{
-		{
-			"Valid via file with suffix - success",
-			"../../testing/fixtures/validator/valid_viafile.cfg",
-			"example.com",
-			"dade.murphy@example.com",
-			"password",
-			nil,
-		},
-		{
-			"Valid via file without suffix - success",
-			"../../testing/fixtures/validator/valid_viafile.cfg",
-			"",
-			"dade.murphy",
-			"password",
-			nil,
-		},
-		{
-			"Invalid via file - failure",
-			"../../testing/fixtures/validator/invalid_viafile.cfg",
-			"",
-			"dade.murphy",
-			"password",
-			fmt.Errorf("Invalid via-file"),
-		},
-		{
-			"Invalid username in via file - failure",
-			"../../testing/fixtures/validator/invalid_username_viafile.cfg",
-			"",
-			"dade.murphy*",
-			"password",
-			fmt.Errorf("Invalid CN or username format"),
-		},
-		{
-			"Missing via file - failure",
-			"MISSING",
-			"",
-			"dade.murphy",
-			"password",
-			fmt.Errorf("stat MISSING: no such file or directory"),
-		},
-		{
-			"Via file is a dir - failure",
-			"../../testing/fixtures/validator/",
-			"",
-			"dade.murphy",
-			"password",
-			fmt.Errorf("read ../../testing/fixtures/validator/: is a directory"),
-		},
-	}
-	for _, test := range tests {
-		t.Run(test.testName, func(t *testing.T) {
-			v := NewOktaOpenVPNValidator()
-			v.api = oktaApiAuth.NewOktaApiAuth()
-			v.api.ApiConfig.UsernameSuffix = test.usernameSuffix
-			err := v.LoadViaFile(test.path)
-			if test.err == nil {
-				assert.Nil(t, err)
-				assert.NotNil(t, v.api.UserConfig)
-				assert.Equal(t, test.expectedUsername, v.api.UserConfig.Username)
-				assert.Equal(t, test.expectedPassword, v.api.UserConfig.Password)
-			} else {
-				assert.Equal(t, test.err.Error(), err.Error())
-			}
-		})
-	}
-}
-
-func setEnv(e map[string]string) {
-	for k, v := range e {
-		os.Setenv(k, v)
-	}
-}
-
-func unsetEnv(e map[string]string) {
-	for k := range e {
-		os.Unsetenv(k)
-	}
-}
-
-func TestLoadEnvVars(t *testing.T) {
-	tests := []testEnvVar{
-		{
-			"Test username/allowUntrustedUsers/usernameSuffix - succes",
-			"example.com",
-			true,
-			true,
-			"dade.murphy@example.com",
-			map[string]string{
-				"username":     "dade.murphy",
-				"common_name":  "",
-				"password":     "password",
-				"untrusted_ip": "1.2.3.4",
-			},
-			nil,
-		},
-		{
-			"Test username/no password - failure",
-			"example.com",
-			true,
-			false,
-			"dade.murphy@example.com",
-			map[string]string{
-				"username":     "dade.murphy",
-				"common_name":  "",
-				"password":     "",
-				"untrusted_ip": "1.2.3.4",
-			},
-			fmt.Errorf("No password"),
-		},
-		{
-			"Test username/!allowUntrustedUsers/usernameSuffix - success",
-			"example.com",
-			false,
-			false,
-			"dade.murphy@example.com",
-			map[string]string{
-				"username":     "dade.murphy",
-				"common_name":  "",
-				"password":     "password",
-				"untrusted_ip": "1.2.3.4",
-			},
-			nil,
-		},
-		{
-			"Test common_name/!allowUntrustedUsers/usernameSuffix - success",
-			"example.com",
-			false,
-			true,
-			"dade.murphy2@example.com",
-			map[string]string{
-				"username":     "dade.murphy",
-				"common_name":  "dade.murphy2",
-				"password":     "password",
-				"untrusted_ip": "1.2.3.4",
-			},
-			nil,
-		},
-		{
-			"Test username/common_name/allowUntrustedUsers/usernameSuffix - success",
-			"example.com",
-			true,
-			true,
-			"dade.murphy@example.com",
-			map[string]string{
-				"username":     "dade.murphy",
-				"common_name":  "dade.murphy2",
-				"password":     "password",
-				"untrusted_ip": "1.2.3.4",
-			},
-			nil,
-		},
-		{
-			"Test empty username/common_name - failure",
-			"example.com",
-			false,
-			false,
-			"dade.murphy@example.com",
-			map[string]string{
-				"username":     "",
-				"common_name":  "",
-				"password":     "password",
-				"untrusted_ip": "1.2.3.4",
-			},
-			fmt.Errorf("No CN or username"),
-		},
-		{
-			"Test invalid username/common_name - failure",
-			"example.com",
-			false,
-			false,
-			"dade.murphy@example.com",
-			map[string]string{
-				"username":     "dade.murphy*",
-				"common_name":  "",
-				"password":     "password",
-				"untrusted_ip": "1.2.3.4",
-			},
-			fmt.Errorf("Invalid CN or username format"),
-		},
-	}
-	for _, test := range tests {
-		t.Run(test.testName, func(t *testing.T) {
-			setEnv(test.env)
-			v := NewOktaOpenVPNValidator()
-			v.api.ApiConfig.UsernameSuffix = test.usernameSuffix
-			v.api.ApiConfig.AllowUntrustedUsers = test.allowUntrustedUsers
-			err := v.LoadEnvVars(nil)
-			unsetEnv(test.env)
-			assert.Equal(t, test.expectedTrusted, v.usernameTrusted)
-			if test.err == nil {
-				assert.Nil(t, err)
-				assert.Equal(t, test.expectedUsername, v.api.UserConfig.Username)
-			} else {
-				assert.Equal(t, test.err.Error(), err.Error())
-			}
-		})
-	}
-}
 
 // Authenticate() already done in oktaApiAuth and Load*
